rename box2 box3 projectToPlane
