+++ lib/box2/index.js
@@ -26,8 +26,8 @@ exports.getPoints = exports.getPath
 exports.getPolygon = exports.getPath
 exports.getSize = require('./getSize')
 exports.homothety = require('./homothety')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.resizeBy = require('./resizeBy')
 exports.resizeTo = require('./resizeTo')
 exports.rotateBy = require('./rotateBy')
