+++ lib/box3/index.js
@@ -21,8 +21,8 @@ exports.getMinimumBounds = require('./getMinimumBounds')
 exports.getSize = require('./getSize')
 exports.getVolume = require('./getVolume')
 exports.homothety = require('./homothety')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.scaleBy = exports.homothety
 exports.transitFrom = require('./transitFrom')
 exports.transitTo = require('./transitTo')
