+++ lib/box3/projectToPlane.js
